# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Movimento 7 API
  version: 0.1.0
  description: API versionada da plataforma Movimento 7.
servers:
  - url: https://movimento7.com.br/api/v1
    description: Produção via proxy same-origin
...
paths:
...
  /admin/editions:
    get:
      tags: [admin]
      operationId: listAdminEditions
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [admin]
      operationId: createAdminEdition
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminEditionInput' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /admin/editions/{id}:
    patch:
      tags: [admin]
      operationId: updateAdminEdition
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminEditionInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /admin/editions/{id}/status:
    patch:
      tags: [admin]
      operationId: updateAdminEditionStatus
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [draft, published, closed, archived] }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
components:
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
  schemas:
    Envelope:
      type: object
      required: [data, meta, error, request_id]
      properties:
        data: {}
        meta: { type: object, additionalProperties: true }
        error:
          oneOf: [{ type: 'null' }, { $ref: '#/components/schemas/ApiError' }]
        request_id: { type: [string, 'null'] }
    ApiError:
      type: object
      required: [code, message, fields]
      properties:
        code: { type: string }
        message: { type: string }
        fields:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    AdminEditionInput:
      type: object
      required: [name, slug, starts_at, ends_at, registration_opens_at, registration_closes_at]
      properties:
        name: { type: string, minLength: 2, maxLength: 140 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 160 }
        description: { type: [string, 'null'], maxLength: 5000 }
        starts_at: { type: [string, 'null'], format: date-time }
        ends_at: { type: [string, 'null'], format: date-time }
        registration_opens_at: { type: [string, 'null'], format: date-time }
        registration_closes_at: { type: [string, 'null'], format: date-time }
        location: { type: [string, 'null'], maxLength: 180 }
        address: { type: [string, 'null'], maxLength: 300 }
        map_url: { type: [string, 'null'], format: uri }
        capacity: { type: [integer, 'null'], minimum: 1, maximum: 100000 }
        retention_days: { type: integer, minimum: 30, maximum: 3650 }
  responses:
    Success:
      description: Resposta bem-sucedida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Error:
      description: Erro esperado e seguro
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }